// src/hooks/snacks/type.ts
export type FoodList = {
  id: number;
  productName: string;
  productImage: string;
  productDescription: string;
  productQuantity: number;
  productPrice: number;
  productTax: number;
  productSize: string;
  productType: string;
  category: string;
  subCategory: string;
  productAddon: number;
  productExtra: any;
};